feat: Refactor Carousel component to support items and labels, and update Blog and ArticleBatch components for responsive slider mode
