+++ apps/website/src/components/Pages/HomePage/Showcase/Showcase.tsx
@@ -1,13 +1,41 @@
 "use client";
 
-import { Bars, Carousel, useClientInfoService } from "@burneeble/ui-components";
+import {
+  Bars,
+  Carousel,
+  CTA,
+  useClientInfoService,
+} from "@burneeble/ui-components";
 import { ShowcaseProps } from "./Showcase.types";
 import { useRouter } from "next/navigation";
+import { cn } from "@/lib/utils";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faArrowUpRightFromSquare } from "@fortawesome/free-solid-svg-icons";
+import { useEffect, useRef, useState } from "react";
 
 const Showcase = (props: ShowcaseProps) => {
+  //States
+  const [showHover, setShowHover] = useState<boolean>(false);
+
   //Hooks
   const { screen } = useClientInfoService();
   const router = useRouter();
+  const hoverLayer = useRef<HTMLDivElement>(null);
+
+  //Effects
+  useEffect(() => {
+    const handleClickOutside = (event: any) => {
+      if (hoverLayer.current && !hoverLayer.current.contains(event.target)) {
+        setShowHover(false);
+      }
+    };
+
+    if (screen === "sm" || screen === "md")
+      document.addEventListener("mousedown", handleClickOutside);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [hoverLayer, screen]);
 
   return (
     <section
@@ -65,7 +93,120 @@ const Showcase = (props: ShowcaseProps) => {
       </div>
       <div className={`carousel-wrapper tw-w-full tw-max-w-[1200px]`}>
         <Carousel
-          projects={props.projects}
+          labels={props.projects.map((proj) => {
+            return proj.categories.length <= 3
+              ? proj.categories
+              : proj.categories.slice(0, 3);
+          })}
+          items={props.projects.map((proj, i) => {
+            return (
+              <div
+                key={i}
+                className={`
+                  carousel-image-wrapper tw-relative tw-flex
+                  tw-aspect-[1920/1080] tw-w-full tw-items-center
+                  tw-justify-center tw-border-4 tw-border-solid
+                  tw-border-primary tw-bg-black
+                `}
+              >
+                <div
+                  className={cn(
+                    `
+                      hover-layer tw-absolute tw-left-0 tw-top-0 tw-h-full
+                      tw-w-full tw-bg-black/60 tw-opacity-0 tw-backdrop-blur-sm
+                      tw-transition-all tw-duration-500 tw-ease-in-out
+
+                      hover:tw-opacity-100
+                    `,
+                    (screen === "sm" || screen === "md") &&
+                      showHover &&
+                      "tw-opacity-100"
+                  )}
+                  onClick={() => {
+                    if (screen === "sm" || screen === "md") {
+                      setShowHover(true);
+                    }
+                  }}
+                  ref={hoverLayer}
+                >
+                  <div
+                    className={cn(
+                      `
+                        content tw-mx-auto tw-flex tw-h-full tw-w-full
+                        tw-flex-col tw-items-center tw-justify-center
+                        tw-gap-[10px]
+
+                        md:tw-gap-[20px]
+                      `,
+                      screen === "sm" || screen === "md"
+                        ? "tw-max-w-[95%]"
+                        : "tw-max-w-[80%]"
+                    )}
+                  >
+                    <h2
+                      className={`
+                        project-title tw-flex tw-items-center tw-gap-[10px]
+                        tw-text-center tw-font-bowlby-one tw-text-2xl
+                        tw-font-normal tw-text-headings
+
+                        lg:tw-text-5xl
+
+                        md:tw-text-4xl
+                      `}
+                    >
+                      {proj.title}
+                      <FontAwesomeIcon
+                        onClick={() => {
+                          window.open(proj.projectUrl, "_blank");
+                        }}
+                        icon={faArrowUpRightFromSquare}
+                        className={`
+                          tw-cursor-pointer tw-text-xl tw-text-primary
+                          tw-transition-all tw-duration-200 tw-ease-in-out
+
+                          hover:tw-brightness-125 hover:tw-filter
+
+                          lg:tw-text-4xl
+
+                          md:tw-text-3xl
+                        `}
+                      />
+                    </h2>
+                    <p
+                      className={`
+                        desc tw- tw-text-center tw-font-inter tw-text-md
+                        tw-font-normal tw-text-body
+
+                        lg:tw-text-2xl
+
+                        md:tw-text-lg md:tw-leading-[35px]
+                      `}
+                    >
+                      {proj.description}
+                    </p>
+                    <CTA
+                      projectUrl={`/project/${proj.title
+                        .toLowerCase()
+                        .replace(/[^a-zA-Z0-9\s]/g, "")
+                        .replaceAll(" ", "-")}`}
+                      text={"View Details"}
+                      size={
+                        screen === "sm" || screen === "md" ? "sm" : "default"
+                      }
+                    />
+                  </div>
+                </div>
+                <img
+                  src={proj.thumbnail}
+                  alt={""}
+                  width={1920}
+                  height={1080}
+                  className="carousel-image tw-h-full tw-w-auto"
+                  loading="lazy"
+                />
+              </div>
+            );
+          })}
           cta={{
             children: "See All Projects",
             variant: "secondary",
