+++ packages/ui-components/src/components/widgets/Carousel/Carousel.types.ts
@@ -1,14 +1,16 @@
 import { ButtonProps } from "@/components/ui";
 
-export interface Project {
-  thumbnail: string;
-  categories: string[];
-  description: string;
-  title: string;
-  projectUrl: string;
-}
+// export interface Project {
+//   thumbnail: string;
+//   categories: string[];
+//   description: string;
+//   title: string;
+//   projectUrl: string;
+// }
 
 export interface CarouselProps {
-  projects: Project[];
+  // projects: Project[];
+  items: JSX.Element[];
+  labels?: string[][];
   cta?: ButtonProps;
 }
